Split onlyChanged.test.js into 2 suits
